+++ packages/shared/lib/transports.dart
@@ -2,5 +2,6 @@
 library transports;
 
 export 'src/transports/polling/polling.dart' show PollingTransport;
+export 'src/transports/websocket/websocket.dart' show WebSocketTransport;
 export 'src/transports/connection_type.dart' show ConnectionType;
 export 'src/transports/transport.dart' show Transport;